Added Sparkle.framework. Cleaned up code.
Added Sparkle.framework to enable auto-updating for users. Because we
expect to add more features. And sentence fragments. Take that. English
teachers.

Also, got rid of almost exactly half of the warnings in the code. The
other half are pointers to places where we can improve the application.

Why am I saying _we_ all of the time in here? I’m the only one making
changes at the moment. Yes, I suppose this is in anticipation of your
pull requests. We already worked together on a pull request. It was
amazing. We open-sourced it up. It was unbelievably thrilling.

We can do it again. I believe in you. Or something like that. Be
motivated. Go look at a picture of a tree.