README.md: don’t encourage turning gatekeeper off
Gatekeeper isn’t there to annoy us. It is a real security feature with real proven benefits.

The type of user that needs to be informed about Gatekeeper and why the app is not running is precisely the type of user that should not turn Gatekeeper off. Encouraging users to do so without a proper understanding of the consequences is doing them a disservice.